# backend/docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  piemdm_internal_model.Permission:
    properties:
      action:
        description: 操作类型,如 read, write
        type: string
      children:
        description: 关联关系
        items:
          $ref: '#/definitions/piemdm_internal_model.Permission'
        type: array
      code:
        description: 权限标识,如 user:list
        type: string
      createdAt:
        type: string
      createdBy:
        description: 创建人
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 描述
        type: string
      id:
        type: integer
      name:
        description: 权限名称
        type: string
      parent_id:
        description: 父权限ID,用于构建树状结构
        type: integer
      resource:
        description: 资源类型,如 user, role
        type: string
      status:
        description: 状态：Normal 正常 Frozen 已冻结 Deleted 已删除
        type: string
      updatedAt:
        type: string
      updatedBy:
        description: 更新人
        type: string
    required:
    - code
    - name
    type: object
  piemdm_internal_model.Role:
    properties:
      code:
        description: 表名,英文
        type: string
      createdAt:
        type: string
      createdBy:
        description: 创建人
        type: string
      data_scope:
        description: '数据范围: All, Subordinate, Self'
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      desc:
        description: 描述
        type: string
      id:
        type: integer
      name:
        description: 名称，可以中文
        type: string
      permissions:
        items:
          $ref: '#/definitions/piemdm_internal_model.Permission'
        type: array
      status:
        description: 状态：Normal 正常 Frozen 已冻结 Deleted 已删除
        type: string
      updatedAt:
        type: string
      updatedBy:
        description: 更新人
        type: string
    required:
    - code
    - name
    type: object
...
  piemdm_internal_model.User:
    properties:
      admin:
        description: 是否为管理员
        maxLength: 3
        type: string
      avatar:
        description: 头像地址
        maxLength: 255
        type: string
      createdBy:
        description: 创建人
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 简介
        maxLength: 255
        type: string
      displayName:
        description: 显示名称
        maxLength: 64
        type: string
      email:
        description: 邮箱
        maxLength: 255
        type: string
      employeeID:
        description: 工号
        maxLength: 64
        type: string
      firstName:
        description: 名称
        maxLength: 64
        type: string
      id:
        type: integer
      language:
        description: 语言
        maxLength: 8
        type: string
      lastLogin:
        description: 最后登录时间
        type: string
      lastName:
        description: 姓氏
        maxLength: 64
        type: string
      password:
        description: 密码
        maxLength: 128
        type: string
      permissions:
        description: 权限列表(不持久化)
        items:
          type: string
        type: array
      phone:
        description: 电话
        maxLength: 64
        type: string
      roles:
        items:
          $ref: '#/definitions/piemdm_internal_model.Role'
        type: array
      sex:
        description: 性别
        maxLength: 8
        type: string
      status:
        description: |-
          状态：Normal 正常 Frozen 已冻结 Deleted 已删除
          状态：Normal 正常 Frozen 已冻结 Deleted 已删除
        type: string
      superiorUsername:
        description: 上级用户名
        maxLength: 64
        type: string
      updatedAt:
        type: string
      updatedBy:
        description: 更新人
        type: string
      username:
        description: 用户名
        maxLength: 64
        type: string
    required:
    - admin
    - displayName
    - email
    - employeeID
    - password
    - superiorUsername
    - username
    type: object
...
  piemdm_internal_pkg_request.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
host: localhost:8787
info:
  contact:
    email: support@piemdm.com
    name: API Support
  description: PieMDM Master Data Management System API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PieMDM API
  version: "1.0"
paths:
...
  /admin/users:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 15
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/piemdm_internal_model.User'
            type: array
      summary: 获取用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.User'
      summary: 创建用户
      tags:
      - 用户管理
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 删除用户
      tags:
      - 用户管理
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.User'
      summary: 获取用户详情
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 更新用户
      tags:
      - 用户管理
  /admin/users/{id}/roles:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/piemdm_internal_model.Role'
            type: array
      summary: 获取用户角色
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色ID列表
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 更新用户角色
      tags:
      - 用户管理
  /admin/users/batch:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 批量删除用户
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 批量更新请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 批量更新用户状态
      tags:
      - 用户管理
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/piemdm_internal_pkg_request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - 用户管理
  /auth/validate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 验证Token
      tags:
      - 用户管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"